[VIRT] Quarantine post-copy migration tests on RHCOS 10
Post-copy migration fails on clusters running RHCOS 10 or above.
Move the RHCOS version detection and bug-check fixtures from
tests/observability/conftest.py to tests/conftest.py so they can
be shared across teams, and add a module-scoped xfail guard that
skips all post-copy migration tests before creating any resources
when the bug is active.
