+++ tests/conftest.py
@@ -65,7 +65,7 @@ from ocp_resources.virtual_machine_instance_migration import (
 from ocp_resources.virtual_machine_instancetype import VirtualMachineInstancetype
 from ocp_resources.virtual_machine_preference import VirtualMachinePreference
 from ocp_utilities.monitoring import Prometheus
-from packaging.version import parse
+from packaging.version import Version, parse
 from pytest_testconfig import config as py_config
 from timeout_sampler import TimeoutSampler
 
@@ -175,6 +175,7 @@ from utilities.infra import (
     run_virtctl_command,
     scale_deployment_replicas,
 )
+from utilities.jira import is_jira_open
 from utilities.network import (
     EthernetNetworkConfigurationPolicy,
     MacPool,
@@ -498,6 +499,33 @@ def workers(nodes):
 @pytest.fixture(scope="session")
 def control_plane_nodes(nodes):
     return get_nodes_with_label(nodes=nodes, label=f"{NODE_ROLE_KUBERNETES_IO}/control-plane")
+
+
+@pytest.fixture(scope="session")
+def workers_rhcos_version(schedulable_nodes):
+    """Returns a dict mapping each schedulable node name to its RHCOS version.
+
+    Returns:
+        dict[str, str]: Node name to RHCOS version (e.g. {"node-1": "10.2.20260408", ...}).
+    """
+    rhcos_version_re = re.compile(r"CoreOS\s+([\d.]+)")
+    versions = {}
+    for node in schedulable_nodes:
+        os_image = node.instance.status.nodeInfo.osImage
+        match = rhcos_version_re.search(string=os_image)
+        assert match, f"Failed to parse RHCOS version from osImage '{os_image}' on node '{node.name}'"
+        versions[node.name] = match.group(1)
+    return versions
+
+
+@pytest.fixture(scope="session")
+def cluster_has_rhcos10_or_above(workers_rhcos_version):
+    return any(Version(ver) >= Version("10") for ver in workers_rhcos_version.values())
+
+
+@pytest.fixture(scope="session")
+def is_postcopy_migration_bug_open(cluster_has_rhcos10_or_above):
+    return cluster_has_rhcos10_or_above and is_jira_open(jira_id="CNV-84023")
 
 
 @pytest.fixture(scope="session")
